+++ crates/nose-cli/src/verify_admission/runtime_boundary/tests/async_runtime.rs
@@ -9,6 +9,7 @@ use super::super::runtime_boundary_missing_evidence_with_context;
 
 mod imported_bindings;
 mod java;
+mod python;
 mod rust_block_on;
 mod rust_block_on_root_shadow;
 mod rust_block_on_self_field;